Move back patch file in project rooy directory
The patch file needs to be in the project root directory in order to be loaded for the rpm generation.
The patch file needs also to be into the devontainer build directory; this means that, at this moment, we use a copy of the patch file but an alternative solution can be find.