Merge all 'yum install' commands in a single layer
To create the image defined in this dockerfile, we've merged three differend images:
 * storm2/build/base
 * storm2/build/ngx-voms-build [issue-17]
 * storm2/ngx_http_voms_module/.devcontainer
Each image has this own 'yum install' command with the packages needed for that layer.

With this commit, all the packages are installed in a single layer to reduce the image dimension.
All the 'yum install' commands are grouped in the setup.sh script.