Rmove the patch file from the .devcontainer folder
The patch file used by the .devcontainer is needed only by  the build-install-ngx-voms.sh script. This script could use the patch file in the project home dir instead of the copy in the container.

Add an echo in the container specifing that the patch file must exists in the directory where the user runs the script.

Add  a check in the script: if the patch file isn't found, the script adivses the user and exits with an error.