Adapt nginx configuration
The nginx used in this branch is the last stable one and not a version compiled by us. To run it inside a container without the root privileges, we have to adapt some configuration in the nginx.conf file. See https://hub.docker.com/_/nginx at section "Running nginx as a non-root user"